+++ src/react/components/picklist/private/PickitemMultiselect.js
@@ -26,6 +26,7 @@ const PickitemMultiselect = ({
   isSelected,
   children,
   checkboxClassName,
+  checkboxRef,
   qaHook,
   primaryElement,
   secondaryElement,
@@ -70,6 +71,7 @@ const PickitemMultiselect = ({
       role="presentation"
     >
       <XUICheckbox
+        checkboxElementClassName={`${pickitemClassName}--input`}
         className={checkboxClasses}
         htmlClassName={`${pickitemClassName}--input`}
         isChecked={isSelected}
@@ -77,6 +79,7 @@ const PickitemMultiselect = ({
         labelClassName={labelClasses}
         onChange={NOOP}
         qaHook={qaHook && `${qaHook}--checkbox`}
+        ref={checkboxRef}
         size={sizeShift('medium', -1)}
         tabIndex={-1}
       >
@@ -96,6 +99,7 @@ const PickitemMultiselect = ({
 
 PickitemMultiselect.propTypes = {
   checkboxClassName: PropTypes.string,
+  checkboxRef: PropTypes.object,
   children: PropTypes.node,
   /** The disabled behaviour and styles are applied when this is true. */
   isDisabled: PropTypes.bool,